cosmetic and diogonal movement set

+++ src/entity/entity.hpp
@@ -4,10 +4,10 @@
 class Entity {
 public:
     virtual ~Entity() {};
-    virtual void move(glm::vec2 direction) = 0;
+    virtual void move(glm::ivec2 direction) = 0;
     virtual void attack() = 0;
 
-    glm::vec2 get_position() { return position; }
+    glm::ivec2 get_position() { return position; }
 protected:
-    glm::vec2 position;
+    glm::ivec2 position;
 };
