+++ src/entity/player.cpp
@@ -2,13 +2,15 @@
 #include <Windows.h>
 #include "player.hpp"
 
-Player::Player() {
+Player::Player()
+{
     position = glm::vec2(DEFAULT_POSITION);
     speed = glm::vec2(DEFAULT_SPEED);
     jump_power = DEFAULT_JUMP_POWER;
 }
 
-Player::Player(int posX, int posY) {
+Player::Player(int posX, int posY)
+{
     position = glm::vec2(posX, posY);
     speed = glm::vec2(DEFAULT_SPEED);
     jump_power = DEFAULT_JUMP_POWER;
@@ -16,17 +18,20 @@ Player::Player(int posX, int posY) {
 
 Player::~Player() {}
 
-void Player::draw() {
+void Player::draw()
+{
     HANDLE hc = GetStdHandle(STD_OUTPUT_HANDLE);
     SetConsoleTextAttribute(hc, 0x06);
     std::cout << texture;
     SetConsoleTextAttribute(hc, 0x07);
 }
 
-void Player::move(glm::vec2 direction) {
+void Player::move(glm::ivec2 direction)
+{
     position += direction;
 }
 
-void Player::attack() {
+void Player::attack()
+{
     std::cout << "Player attacked!\n";
 }
