+++ src/entity/target.hpp
@@ -2,14 +2,15 @@
 #include <glm/glm.hpp>
 #include "entity.hpp"
 
-class Target : public Entity {
+class Target : public Entity
+{
 public:
     bool is_visited = false;
     
     Target(int posX, int posY);
     ~Target() override;
 
-    void move(glm::vec2 direction) override;
+    void move(glm::ivec2 direction) override;
     void attack() override;
     void draw();
